Move lock validation functionality into lock method.
